+++ app/src/main/java/fr/imacaron/groupe19/urgan/common/MostPlayedGame.kt
@@ -7,7 +7,7 @@ data class MostPlayedGame(
     val rank : Int,
 
     @SerializedName("appid")
-    val app_id : Int,
+    val app_id : Long,
 
     val last_week_rank : Int,
